* Modify service bus partition key sample.
