--- servicebus/azure-spring-cloud-stream-binder-servicebus-queue/servicebus-queue-binder-partition-key/README.md
@@ -1,193 +1,0 @@
----
-page_type: sample
-languages:
-- java
-  products:
-- azure-service-bus
-  description: "Azure Spring Cloud Stream Binder Sample project for Service Bus queue client library"
-  urlFragment: "azure-spring-cloud-stream-binder-service-bus-partition-key"
----
-
-# Spring Cloud Azure Stream Binder for Service Bus queue Sample shared library for Java
-
-## Key concepts
-
-This code example demonstrates how to use Spring Cloud Stream binder
-Azure Service Bus queue. 
-The sample application demonstrates how to set the session id, partition key and the priority relationship between them in the message header.
-
-## Getting started
-
-Running this sample will be charged by Azure. You can check the usage and bill at
-[this link][azure-account].
-
-
-### Create Azure resources
-
-We have several ways to config the Spring Cloud Stream Binder for Service
-Bus Queue. You can choose anyone of them.
-
-#### Method 1: Connection string based usage
-
-1.  Create Azure Service Bus namespace and queue.
-    Please see [how to create][create-service-bus].
-
-1.  Update [application.yaml][application.yaml].
-    ```yaml
-    spring:
-      cloud:
-        azure:
-          servicebus:
-            connection-string: [servicebus-namespace-connection-string] 
-        stream:
-          function:
-            definition: consume;supply
-          bindings: 
-            consume-in-0: 
-              destination: [servicebus-queue-name]
-            supply-out-0:
-              destination: [servicebus-queue-name-same-as-above]
-          poller:
-            fixed-delay: 1000
-            initial-delay: 0
-    ```
-
-#### Method 2: Service principal based usage
-
-1.  Create a service principal for use in by your app. Please follow
-    [create service principal from Azure CLI][create-sp-using-azure-cli].
-
-1.  Create Azure Service Bus namespace and queue.
-    Please see [how to create][create-service-bus].
-
-1.  Add Role Assignment for Service Bus. See
-    [Service principal for Azure resources with Service Bus][role-assignment]
-    to add role assignment for Service Bus. Assign `Contributor` role for service bus.
-
-1.  Update [application-sp.yaml][application-sp.yaml].
-    ```yaml
-    spring:
-      cloud:
-        azure:
-          client-id: [service-principal-id]
-          client-secret: [service-principal-secret]
-          tenant-id: [tenant-id]
-          resource-group: [resource-group]
-          servicebus:
-            namespace: [servicebus-namespace]
-        stream:
-          function:
-            definition: consume;supply
-          bindings:
-            consume-in-0:
-              destination: [servicebus-queue-name]
-            supply-out-0:
-              destination: [servicebus-queue-name-same-as-above]
-          poller:
-            fixed-delay: 1000
-            initial-delay: 0
-    ```
-#### Method 3: MSI credential based usage
-
-##### Set up managed identity
-
-Please follow [create managed identity][create-managed-identity] to set up managed identity.
-
-##### Create other Azure resources
-
-1.  Create Azure Service Bus namespace and queue.
-    Please see [how to create][create-service-bus].
-
-1.  Add Role Assignment for Service Bus. See
-    [Managed identities for Azure resources with Service Bus][role-assignment]
-    to add role assignment for Service Bus. Assign `Contributor` role for managed identity.
-
-
-##### Update MSI related properties
-
-1.  Update [application-mi.yaml][application-mi.yaml].
-    ```yaml
-    spring:
-      cloud:
-        azure:
-          msi-enabled: true
-          client-id: [the-id-of-managed-identity]
-          resource-group: [resource-group]
-          subscription-id: [subscription-id]
-          servicebus:
-            namespace: [servicebus-namespace]
-        stream:
-          function:
-            definition: consume;supply
-          bindings:
-            consume-in-0:
-              destination: [servicebus-queue-name]
-            supply-out-0:
-              destination: [servicebus-queue-name-same-as-above]
-          poller:
-            fixed-delay: 1000
-            initial-delay: 0
-    ```
-    > We should specify `spring.profiles.active=mi` to run the Spring Boot application.
-    For App Service, please add a configuration entry for this.
-
-##### Redeploy Application
-
-If you update the `spring.cloud.azure.managed-identity.client-id`
-property after deploying the app, or update the role assignment for
-services, please try to redeploy the app again.
-
-> You can follow
-> [Deploy a Spring Boot JAR file to Azure App Service][deploy-spring-boot-application-to-app-service]
-> to deploy this application to App Service
-
-#### Enable auto create
-
-If you want to auto create the Azure Service Bus instances, make sure you add such properties
-(only support the service principal and managed identity cases):
-
-```yaml
-spring:
-  cloud:
-    azure:
-      subscription-id: [subscription-id]
-      auto-create-resources: true
-      environment: Azure
-      region: [region]
-```
-
-
-## Examples
-
-1. Run the `mvn spring-boot:run` in the root of the code sample to get the app running.
-
-2. Send a POST request
-
-        $ curl -X POST http://localhost:8080/setSessionId?message=hello
-
-3. It will be visible on Azure Portal that sent messages are configured with session id and partition key.
-![Azure portal displays the session id.](docs/image-set-session-id.png )
-
-4. Delete the resources on [Azure Portal][azure-portal] to avoid unexpected charges.
-
-## Troubleshooting
-
-## Next steps
-
-## Contributing
-
-
-<!-- LINKS -->
-
-[azure-account]: https://azure.microsoft.com/account/
-[azure-portal]: https://ms.portal.azure.com/
-[create-service-bus]: https://docs.microsoft.com/azure/service-bus-messaging/service-bus-create-namespace-portal
-[create-azure-storage]: https://docs.microsoft.com/azure/storage/
-[create-sp-using-azure-cli]: https://github.com/Azure-Samples/azure-spring-boot-samples/blob/main/create-sp-using-azure-cli.md
-[create-managed-identity]: https://github.com/Azure-Samples/azure-spring-boot-samples/blob/main/create-managed-identity.md
-[deploy-spring-boot-application-to-app-service]: https://docs.microsoft.com/java/azure/spring-framework/deploy-spring-boot-java-app-with-maven-plugin?toc=%2Fazure%2Fapp-service%2Fcontainers%2Ftoc.json&view=azure-java-stable
-
-[role-assignment]: https://docs.microsoft.com/azure/role-based-access-control/role-assignments-portal
-[application-mi.yaml]: https://github.com/Azure-Samples/azure-spring-boot-samples/blob/main/servicebus/azure-spring-cloud-sample-servicebus-queue-binder/src/main/resources/application-mi.yaml
-[application-sp.yaml]: https://github.com/Azure-Samples/azure-spring-boot-samples/blob/main/servicebus/azure-spring-cloud-sample-servicebus-queue-binder/src/main/resources/application-sp.yaml
-[application.yaml]: https://github.com/Azure-Samples/azure-spring-boot-samples/blob/main/servicebus/azure-spring-cloud-sample-servicebus-queue-binder/src/main/resources/application.yaml
