--- servicebus/azure-spring-cloud-stream-binder-servicebus-queue/servicebus-queue-binder-partition-key/src/main/java/com/azure/spring/sample/servicebus/queue/binder/partitionkey/ServiceBusBinderPartitionKeyApplication.java
@@ -1,12 +1,0 @@
-package com.azure.spring.sample.servicebus.queue.binder.partitionkey;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ServiceBusBinderPartitionKeyApplication {
-
-    public static void main(String[] args) {
-		SpringApplication.run(ServiceBusBinderPartitionKeyApplication.class, args);
-	}
-}
