--- servicebus/azure-spring-cloud-stream-binder-servicebus-queue/servicebus-queue-binder-partition-key/src/main/resources/application-mi.yaml
@@ -1,33 +1,0 @@
-spring:
-  cloud:
-    azure:
-      msi-enabled: true
-      client-id: [the-id-of-managed-identity]
-      resource-group: [resource-group]
-      subscription-id: [subscription-id]
-
-#     Uncomment below configurations if you want to enable auto creating resources.
-#
-#      auto-create-resources: true
-#      environment: Azure
-#      region: [region]
-
-      servicebus:
-        namespace: [servicebus-namespace]
-    stream:
-      function:
-        definition: consume;supply
-      bindings:
-        consume-in-0:
-          destination: [servicebus-queue-name]
-        supply-out-0:
-          destination: [servicebus-queue-name-same-as-above]
-      servicebus:
-        queue:
-          bindings:
-            consume-in-0:
-              consumer:
-                checkpoint-mode: MANUAL
-      poller:
-        fixed-delay: 1000
-        initial-delay: 0
