--- servicebus/azure-spring-cloud-stream-binder-servicebus-queue/servicebus-queue-binder-partition-key/src/main/resources/application-sp.yaml
@@ -1,34 +1,0 @@
-spring:
-  cloud:
-    azure:
-      client-id: [service-principal-id]
-      client-secret: [service-principal-secret]
-      tenant-id: [tenant-id]
-      resource-group: [resource-group]
-
-#     Uncomment below configurations if you want to enable auto creating resources.
-#
-#      subscription-id: [subscription-id]
-#      auto-create-resources: true
-#      environment: Azure
-#      region: [region]
-
-      servicebus:
-        namespace: [servicebus-namespace]
-    stream:
-      function:
-        definition: consume;supply
-      bindings:
-        consume-in-0:
-          destination: [servicebus-queue-name]
-        supply-out-0:
-          destination: [servicebus-queue-name-same-as-above]
-      servicebus:
-        queue:
-          bindings:
-            consume-in-0:
-              consumer:
-                checkpoint-mode: MANUAL
-      poller:
-        fixed-delay: 1000
-        initial-delay: 0
